requirements.txt
graphviz
lark 
z3-solver
pytictoc
hiredis
IPython